Event store is now working.

+++ FFF/src/edu/emory/fff/parser/Event.java
@@ -5,7 +5,7 @@ import java.util.Calendar;
 import java.util.Date;
 
 public class Event {
-	private int id;
+	private long id;
 	private String coordinater;
 	private Date notification_date;
 	
@@ -16,11 +16,11 @@ public class Event {
 	private String body;
 	private ArrayList<Highlight> highlights;
 	
-	public int getID() {
+	public long getID() {
 		return id;
 	}
 	
-	public void setID(int id) {
+	public void setID(long id) {
 		this.id = id;
 	}
